refactor(phase3): extract tier handlers from fetch.ts into src/tiers/
Move githubFetch, firecrawlScrape, crawl4aiFetch/pollCrawl4aiTask,
rawFetch/fetchRawHtmlForMetadata into src/tiers/{github,firecrawl,crawl4ai,raw}.ts.
Shared primitives (USER_AGENT, assertPublicUrl, readBoundedText, TierResult) extracted
to src/fetch-utils.ts to avoid circular imports.

fetch.ts drops from 601 to 296 lines; all tier files <130 lines each. Pure code-move
— no behavior changes.
